Move rate_limit decorator
