+++ backend/api/event/views.py
@@ -11,6 +11,7 @@ from api.availability.utils import get_weekday_date
 from api.decorators import (
     api_endpoint,
     check_auth,
+    rate_limit,
     require_auth,
     validate_json_input,
     validate_output,
@@ -36,7 +37,7 @@ from api.event.utils import (
 )
 from api.models import EventDateTimeslot, EventWeekdayTimeslot, UrlCode, UserEvent
 from api.settings import GENERIC_ERR_RESPONSE
-from api.utils import MessageOutputSerializer, format_event_info, rate_limit
+from api.utils import MessageOutputSerializer, format_event_info
 
 logger = logging.getLogger("api")
